Port GHC plugin to 9.10, add compile-time type verification and e2e tests
Port hsqlx-plugin from GHC 9.4 to 9.10.3:
- Update error diagnostics (mkSimpleUnknownDiagnostic)
- Fix HsModule parameterization (HsModule GhcPs)
- Fix ImportDecl fields (XImportDeclPass, ideclImportList)
- Fix AST constructors (HsPar, HsLet arity changes)
- Fix ExpandedThingTc (replaces ExpansionExpr/HsExpanded)
- Fix expression annotations (noExtField vs noAnn)
- Use GHC's TyCon.isTupleTyCon for reliable tuple detection
- Add plugin package to cabal.project

Fix critical verification gap: the typecheck phase now detects
mkStatement calls (rewritten queryFile) by walking the application
spine, so type mismatches are caught even after the parse-phase
rewrite. All 6 HSQLX error codes verified with tests.

Add HsqlxPluginRequired TypeError: queryFile without the plugin gives
a clear compile-time error with instructions, not a runtime crash.
Placeholder mkStatement calls suppress TypeError when plugin is loaded
but cache is missing (letting HSQLX-001/002 fire instead).

Plugin strips queryFile/queryFileAs from user import lists after
rewrite and marks injected imports as implicit, eliminating spurious
-Wunused-imports warnings under -Werror.

Allow Maybe T to match T for parameter types (valid for SQL NULL).

Add error case tests (HSQLX-001 through 006) and e2e test with
positional params, named params, unit result, and no-param queries.

Update example project to use queryFile with the plugin instead of
manual mkStatement calls. Wire watch command to auto-prepare on
file changes. Add named params test SQL file.

+++ example/src/Main.hs
@@ -13,7 +13,7 @@ module Main where
 
 import Data.Aeson (FromJSON, ToJSON, object, (.=))
 import Data.ByteString.Char8 qualified as BS8
-import Data.Int (Int32)
+import Data.Int (Int32, Int64)
 import Data.Text (Text)
 import Data.Time (UTCTime)
 import GHC.Generics (Generic)
@@ -117,7 +117,7 @@ main = do
     get "/posts/recent" $ do
       n <- queryParamMaybe "n" >>= \case
         Just v -> pure v
-        Nothing -> pure 10 :: ActionM Int32
+        Nothing -> pure 10 :: ActionM Int64
       posts <- liftIO $ withResource pool $ \conn ->
         fetchAll conn listRecentPosts n
       json [ object [ "id" .= i, "title" .= t
